+++ lib/macroy_web/views/todo_view.ex
@@ -1,0 +1,3 @@
+defmodule MacroyWeb.TodoView do
+  use MacroyWeb, :view
+end