+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to
 ## [Unreleased]
 ### Added
 - An owner field in the Todo schema which links a todo to a user.
+- A very basic template for Todo index.
 
 ## [1.2.3] - 2019-11-28
 ### Fixed